Evaluate policies individually on individual resources
