+++ app/models/computed_policy.rb
@@ -9,6 +9,7 @@ class ComputedPolicy < ActiveRecord::Base
   scope :delegable, -> { where(delegable: true) }
   scope :on,        -> (resource_type) { where(resource_type: resource_type) }
 
+  include ComputedPolicyConcern
 
   def self.update_user(user)
     # TODO: Run in background
@@ -26,8 +27,17 @@ class ComputedPolicy < ActiveRecord::Base
   end
 
   def self.authorize(action, resource, user, opts={})
-    scope = self.authorize_scope(action, resource, user, opts)
-    return resource.kind_of?(Resource) ? scope.first : scope
+    return resource.kind_of?(Resource)\
+      ? authorize_instance(action, resource, user, opts)\
+      : authorize_scope(action, resource, user, opts)
+  end
+
+  def self.authorize_instance(action, resource, user, opts={})
+    applicable_policies(action, resource, user, opts).includes(:exceptions).reject do |policy|
+      policy.exceptions.any? do |exception|
+        exception.applies_to?(resource, action, opts)
+      end
+    end.empty? ? nil : resource
   end
 
   def self.authorize_scope(action, resource, user, opts={})
@@ -71,60 +81,6 @@ class ComputedPolicy < ActiveRecord::Base
         institution_id: (resource.institution_id if resource.respond_to?(:institution_id)) }
     end
   end
-
-
-  module ComputedPolicyMethods
-
-    def resource_class
-      Resource.resolve(self.resource_type)[0]
-    end
-
-    def arel_filter
-      table = resource_class.arel_table
-      filters = []
-      filters << table[:id].eq(self.resource_id) if self.resource_id
-      filters << table[:institution_id].eq(self.condition_institution_id) if self.condition_institution_id
-      filters << table[:laboratory_id].eq(self.condition_laboratory_id) if self.condition_laboratory_id
-
-      filters += exceptions.map(&:arel_filter).compact.map(&:not) if respond_to?(:exceptions)
-
-      return filters.compact.inject{|agg, filter| agg.and(filter)}
-    end
-
-    def attributes_equal?(p2)
-      self.delegable == p2.delegable && self.computed_attributes == p2.computed_attributes
-    end
-
-    def computed_attributes
-      attrs = {
-        action: self.action,
-        resource_type: self.resource_type,
-        resource_id: self.resource_id,
-        condition_laboratory_id: self.condition_laboratory_id,
-        condition_institution_id: self.condition_institution_id
-      }
-      attrs[:exceptions_attributes] = self.exceptions.map(&:computed_attributes) if self.respond_to?(:exceptions)
-      return attrs
-    end
-
-    def contains(p2)
-      return (self.action.nil? || self.action == p2.action)\
-        && (self.resource_type.nil? || self.resource_type == p2.resource_type)\
-        && (self.resource_id.nil? || self.resource_id == p2.resource_id)\
-        && (self.conditions.keys.all? {|c| self.conditions[c].nil? || self.conditions[c] == p2.conditions[c] })\
-        && (self.delegable || !p2.delegable)
-    end
-
-    def conditions
-      return {
-        laboratory_id: condition_laboratory_id,
-        institution_id: condition_institution_id
-      }
-    end
-
-  end
-
-  include ComputedPolicyMethods
 
 
   class PolicyComputer
