+++ app/models/computed_policy_exception.rb
@@ -2,6 +2,6 @@ class ComputedPolicyException < ActiveRecord::Base
 
   belongs_to :computed_policy, inverse_of: :exceptions
 
-  include ComputedPolicy::ComputedPolicyMethods
+  include ComputedPolicyConcern
 
 end
